Xena - initial scheduling workflow script
