+++ xena/pages/jobs_page.py
@@ -41,7 +41,8 @@ class JobsPage(DiabloPages):
     RUN_HOUSEKEEPING_JOB_BUTTON = (By.ID, 'run-job-house_keeping')
     RUN_KALTURA_JOB_BUTTON = (By.ID, 'run-job-kaltura')
     RUN_EMAILS_JOB_BUTTON = (By.ID, 'run-job-emails')
-    RUN_REMIND_INVITEES_BUTTON = (By.ID, 'run-job-remind_invitees')
+    RUN_REMIND_SCHEDULED_BUTTON = (By.ID, 'run-job-remind_scheduled')
+    RUN_SEMESTER_START_JOB_BUTTON = (By.ID, 'run-job-semester-start')
     RUN_SIS_DATA_REFRESH_JOB_BUTTON = (By.ID, 'run-job-sis_data_refresh')
 
     SEARCH_HISTORY_INPUT = (By.XPATH, '//label[text()="Search History"]/following-sibling::input')
@@ -89,12 +90,19 @@ class JobsPage(DiabloPages):
         self.wait_for_page_and_click(JobsPage.RUN_EMAILS_JOB_BUTTON)
         self.wait_for_most_recent_job_success(AsyncJob.EMAILS)
 
-    def run_remind_invitees_job(self):
-        app.logger.info('Running reminder invites job')
+    def run_remind_scheduled_job(self):
+        app.logger.info('Running remind scheduled job')
         self.scroll_to_top()
         time.sleep(1)
-        self.wait_for_page_and_click(JobsPage.RUN_REMIND_INVITEES_BUTTON)
-        self.wait_for_most_recent_job_success(AsyncJob.REMIND_INVITEES)
+        self.wait_for_page_and_click(JobsPage.RUN_REMIND_SCHEDULED_BUTTON)
+        self.wait_for_most_recent_job_success(AsyncJob.REMIND_SCHEDULED)
+
+    def run_semester_start_job(self):
+        app.logger.info('Running Semester Start job')
+        self.scroll_to_top()
+        time.sleep(1)
+        self.wait_for_page_and_click(JobsPage.RUN_SEMESTER_START_JOB_BUTTON)
+        self.wait_for_most_recent_job_success(AsyncJob.SEMESTER_START)
 
     def run_sis_data_refresh_job(self):
         app.logger.info('Running SIS data refresh job')
@@ -145,14 +153,14 @@ class JobsPage(DiabloPages):
 
     def disable_all_jobs(self):
         for job in AsyncJob:
-            if job == AsyncJob.REMIND_INVITEES:
+            if job == AsyncJob.REMIND_SCHEDULED:
                 app.logger.info(f'Skipping #{job.value}')
             else:
                 self.disable_job(job)
 
     def search_job_history(self, async_job):
-        app.loggerinfo(f'Searching for {async_job.value}')
-        self.wait_for_element_and_type(async_job.value)
+        app.logger.info(f'Searching for {async_job.value}')
+        self.wait_for_element_and_type(JobsPage.SEARCH_HISTORY_INPUT, async_job.value)
 
     @staticmethod
     def job_row_one_locator():
