--- xena/tests/test_unschedule.py
@@ -1,212 +1,0 @@
-"""
-Copyright ©2024. The Regents of the University of California (Regents). All Rights Reserved.
-
-Permission to use, copy, modify, and distribute this software and its documentation
-for educational, research, and not-for-profit purposes, without fee and without a
-signed licensing agreement, is hereby granted, provided that the above copyright
-notice, this paragraph and the following two paragraphs appear in all copies,
-modifications, and distributions.
-
-Contact The Office of Technology Licensing, UC Berkeley, 2150 Shattuck Avenue,
-Suite 510, Berkeley, CA 94720-1620, (510) 643-7201, otl@berkeley.edu,
-http://ipira.berkeley.edu/industry-info for commercial licensing opportunities.
-
-IN NO EVENT SHALL REGENTS BE LIABLE TO ANY PARTY FOR DIRECT, INDIRECT, SPECIAL,
-INCIDENTAL, OR CONSEQUENTIAL DAMAGES, INCLUDING LOST PROFITS, ARISING OUT OF
-THE USE OF THIS SOFTWARE AND ITS DOCUMENTATION, EVEN IF REGENTS HAS BEEN ADVISED
-OF THE POSSIBILITY OF SUCH DAMAGE.
-
-REGENTS SPECIFICALLY DISCLAIMS ANY WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE. THE
-SOFTWARE AND ACCOMPANYING DOCUMENTATION, IF ANY, PROVIDED HEREUNDER IS PROVIDED
-"AS IS". REGENTS HAS NO OBLIGATION TO PROVIDE MAINTENANCE, SUPPORT, UPDATES,
-ENHANCEMENTS, OR MODIFICATIONS.
-"""
-
-import pytest
-from xena.models.publish_type import PublishType
-from xena.models.recording_approval_status import RecordingApprovalStatus
-from xena.models.recording_schedule import RecordingSchedule
-from xena.models.recording_scheduling_status import RecordingSchedulingStatus
-from xena.models.recording_type import RecordingType
-from xena.test_utils import util
-
-
-@pytest.mark.usefixtures('page_objects')
-class TestUnschedule:
-
-    test_data = util.get_test_script_course('test_unschedule')
-    section = util.get_test_section(test_data)
-    recording_schedule = RecordingSchedule(section)
-
-    # DELETE PRE-EXISTING DATA
-
-    def test_disable_jobs(self):
-        self.login_page.load_page()
-        self.login_page.dev_auth()
-        self.ouija_page.click_jobs_link()
-        self.jobs_page.disable_all_jobs()
-
-    def test_delete_old_kaltura_series(self):
-        self.kaltura_page.log_in_via_calnet(self.calnet_page)
-        self.kaltura_page.reset_test_data(self.term, self.recording_schedule)
-
-    def test_delete_old_diablo_data(self):
-        util.reset_sign_up_test_data(self.section)
-        self.recording_schedule.approval_status = RecordingApprovalStatus.NOT_INVITED
-        self.recording_schedule.scheduling_status = RecordingSchedulingStatus.NOT_SCHEDULED
-
-    # ADMIN QUEUES COURSE FOR SCHEDULING AND UN-SCHEDULES PRIOR TO SERIES CREATION IN KALTURA
-
-    def test_admin_approve(self):
-        self.sign_up_page.load_page(self.section)
-        self.sign_up_page.select_rec_type(RecordingType.SCREENCAST.value['option'])
-        self.sign_up_page.select_publish_type(PublishType.BCOURSES.value)
-        self.sign_up_page.click_approve_button()
-
-    def test_admin_queued_for_scheduling(self):
-        self.sign_up_page.wait_for_approvals_msg('This course is currently queued for scheduling')
-
-    def test_unschedule_queued_cancel(self):
-        self.sign_up_page.cancel_unscheduling()
-
-    def test_unschedule_queued(self):
-        self.sign_up_page.confirm_unscheduling(self.recording_schedule)
-
-    def test_unschedule_queued_opted_out(self):
-        assert self.sign_up_page.is_opted_out()
-
-    def test_unschedule_queued_job_run(self):
-        self.ouija_page.click_jobs_link()
-        self.jobs_page.run_kaltura_job()
-        assert not util.get_kaltura_id(self.recording_schedule, self.term)
-
-    # VERIFY FILTERS
-
-    def test_unsched_filter_all(self):
-        self.ouija_page.load_page()
-        self.ouija_page.search_for_course_code(self.section)
-        self.ouija_page.filter_for_all()
-        assert self.ouija_page.is_course_in_results(self.section) is True
-
-    def test_unsched_approval_status(self):
-        visible_status = self.ouija_page.course_row_approval_status_el(self.section).text.strip()
-        assert visible_status == self.recording_schedule.approval_status.value
-
-    def test_unsched_sched_status(self):
-        visible_status = self.ouija_page.course_row_sched_status_el(self.section).text.strip()
-        assert visible_status == self.recording_schedule.scheduling_status.value
-
-    def test_unsched_filter_no_email(self):
-        self.ouija_page.filter_for_do_not_email()
-        assert self.ouija_page.is_course_in_results(self.section) is True
-
-    def test_unsched_filter_not_invited(self):
-        self.ouija_page.filter_for_not_invited()
-        assert self.ouija_page.is_course_in_results(self.section) is False
-
-    def test_unsched_filter_invited(self):
-        self.ouija_page.filter_for_invited()
-        assert self.ouija_page.is_course_in_results(self.section) is False
-
-    def test_unsched_filter_partial_approve(self):
-        self.ouija_page.filter_for_partially_approved()
-        assert self.ouija_page.is_course_in_results(self.section) is False
-
-    def test_unsched_filter_queued(self):
-        self.ouija_page.filter_for_queued_for_scheduling()
-        assert self.ouija_page.is_course_in_results(self.section) is False
-
-    def test_unsched_filter_scheduled(self):
-        self.ouija_page.filter_for_scheduled()
-        assert self.ouija_page.is_course_in_results(self.section) is False
-
-    # INSTRUCTOR SIGNS UP AND RECORDINGS ARE QUEUED FOR SCHEDULING
-
-    def test_hit_sign_up(self):
-        self.ouija_page.load_page()
-        self.ouija_page.log_out()
-        self.login_page.dev_auth(self.section.instructors[0].uid)
-        self.ouija_page.click_sign_up_page_link(self.section)
-
-    def test_sign_up(self):
-        self.sign_up_page.select_rec_type(RecordingType.VIDEO_SANS_OPERATOR.value['option'])
-        self.sign_up_page.select_publish_type(PublishType.BCOURSES.value)
-        self.sign_up_page.click_agree_checkbox()
-        self.sign_up_page.click_approve_button()
-
-    def test_confirmation(self):
-        self.sign_up_page.wait_for_approvals_msg('This course is currently queued for scheduling')
-
-    def test_schedule_recordings(self):
-        self.sign_up_page.log_out()
-        self.login_page.dev_auth()
-        self.ouija_page.click_jobs_link()
-        self.jobs_page.run_kaltura_job()
-        util.get_kaltura_id(self.recording_schedule, self.term)
-
-    # SERIES IS CREATED IN KALTURA
-
-    def test_click_series_link(self):
-        self.sign_up_page.load_page(self.section)
-        self.sign_up_page.click_kaltura_series_link(self.recording_schedule)
-        self.kaltura_page.wait_for_delete_button()
-
-    def test_series_title(self):
-        expected = f'{self.section.code}, {self.section.number} ({self.term.name})'
-        assert self.kaltura_page.visible_series_title() == expected
-
-    def test_close_kaltura_window(self):
-        self.kaltura_page.close_window_and_switch()
-
-    # ADMIN UN-SCHEDULES THE COURSE
-
-    def test_unschedule_confirm(self):
-        self.sign_up_page.confirm_unscheduling(self.recording_schedule)
-
-    def test_opted_out_sign_up(self):
-        assert self.sign_up_page.is_opted_out()
-
-    def test_series_deleted(self):
-        self.kaltura_page.load_event_edit_page(self.recording_schedule.series_id)
-        self.kaltura_page.wait_for_title('Access Denied - UC Berkeley - Test')
-
-    # VERIFY FILTERS
-
-    def test_unsched_again_filter_all(self):
-        self.ouija_page.load_page()
-        self.ouija_page.search_for_course_code(self.section)
-        self.ouija_page.filter_for_all()
-        assert self.ouija_page.is_course_in_results(self.section) is True
-
-    def test_unsched_again_approval_status(self):
-        visible_status = self.ouija_page.course_row_approval_status_el(self.section).text.strip()
-        assert visible_status == self.recording_schedule.approval_status.value
-
-    def test_unsched_again_sched_status(self):
-        visible_status = self.ouija_page.course_row_sched_status_el(self.section).text.strip()
-        assert visible_status == self.recording_schedule.scheduling_status.value
-
-    def test_unsched_again_filter_no_email(self):
-        self.ouija_page.filter_for_do_not_email()
-        assert self.ouija_page.is_course_in_results(self.section) is True
-
-    def test_unsched_again_filter_not_invited(self):
-        self.ouija_page.filter_for_not_invited()
-        assert self.ouija_page.is_course_in_results(self.section) is False
-
-    def test_unsched_again_filter_invited(self):
-        self.ouija_page.filter_for_invited()
-        assert self.ouija_page.is_course_in_results(self.section) is False
-
-    def test_unsched_again_filter_partial_approve(self):
-        self.ouija_page.filter_for_partially_approved()
-        assert self.ouija_page.is_course_in_results(self.section) is False
-
-    def test_unsched_again_filter_queued(self):
-        self.ouija_page.filter_for_queued_for_scheduling()
-        assert self.ouija_page.is_course_in_results(self.section) is False
-
-    def test_unsched_again_filter_scheduled(self):
-        self.ouija_page.filter_for_scheduled()
-        assert self.ouija_page.is_course_in_results(self.section) is False
